Phase 5 (3/3): split runtime, CLI, python crates
nest-runtime/src/lib.rs (572) → split:
  lib.rs            48 — pub re-exports, SearchHit/SearchResult
  mmap_file.rs     245 — MmapNestFile + open + accessors + inspect/revalidate
  search.rs        260 — search/search_ann/search_hybrid + helpers
  materialize.rs    50 — materialize_f32_vectors (dtype-agnostic decode)

nest-runtime/src/simd.rs (432) → simd/{mod, scalar, avx2, neon, tests}.rs
  mod.rs   141 — SimdBackend, dispatch, score_int8_section
  scalar    40 — auto-vectorizable fallback for f32 / f16 / i8
  avx2      59 — x86_64 AVX2+FMA implementations (gated)
  neon      88 — aarch64 NEON implementations (gated)
  tests     94 — SIMD parity vs scalar, deterministic LCG inputs

nest-runtime/src/bm25.rs (377) → bm25/{mod, index, codec, tokenize, fusion, tests}.rs
  mod.rs        40 — re-exports
  index.rs     124 — Bm25Index + build/search
  codec.rs     132 — to_bytes/from_bytes + ByteCursor
  tokenize.rs   29 — Unicode-aware tokenizer
  fusion.rs     30 — rrf_union (RRF k=60)
  tests.rs      52 — 4 tests

nest-cli/src/main.rs (582) → main + cmd/{...}.rs
  main.rs                 121 — Cli struct + Commands enum + dispatch
  cmd/{inspect,validate,
       search,search_ann,
       search_text,
       benchmark,
       cite,stats}.rs    one per subcommand, 20–145 lines each
  cmd/util.rs              66 — print_result, encoding_name, default_embedder_path

nest-python/src/lib.rs (428) → split:
  lib.rs           20 — pymodule registration
  nest_file.rs    167 — NestFile class + SearchHitPy + From impl
  build_fn.rs     232 — build() function + parse_chunks helper
  chunk_id_fn.rs   23 — chunk_id() function

Module rename note: build/chunk_id had to become build_fn/chunk_id_fn
because pyo3's #[pyfunction] expands to a sibling const _PYO3_DEF_<name>
that conflicts with same-named modules.

Final state:
  Largest .rs (non-tests) is search.rs at 260 lines.
  No file in crates/**/src exceeds 300 lines.
  134 tests still pass. clippy clean. fmt clean.
  Python tests + model_hash E2E green. .so rebuilt and tested.
