+++ crates/nest-cli/src/cmd/benchmark.rs
@@ -1,0 +1,93 @@
+//! `nest benchmark <file> -q N -k K [--ann EF]` — latency stats over
+//! random queries, with optional ANN comparison + recall@k vs exact.
+
+use anyhow::Result;
+use std::path::PathBuf;
+
+pub fn run(file: PathBuf, n_queries: usize, k: i32, ann_ef: Option<usize>) -> Result<()> {
+    let runtime = nest_runtime::MmapNestFile::open(&file)?;
+    let dim = runtime.embedding_dim();
+
+    let mut queries = Vec::with_capacity(n_queries);
+    for _ in 0..n_queries {
+        let mut q: Vec<f32> = (0..dim).map(|_| rand::random::<f32>()).collect();
+        let norm = q.iter().map(|x| x * x).sum::<f32>().sqrt();
+        if norm > 0.0 {
+            for x in &mut q {
+                *x /= norm;
+            }
+        }
+        queries.push(q);
+    }
+
+    let exact_times = run_bench(&runtime, &queries, |rt, q| rt.search(q, k))?;
+    println!(
+        "Exact ({} queries, dim={}, dtype={}, simd={}):",
+        n_queries,
+        dim,
+        runtime.dtype().name(),
+        runtime.simd_backend().name()
+    );
+    print_latency(&exact_times);
+
+    if let Some(ef) = ann_ef {
+        if !runtime.has_ann() {
+            println!("(no HNSW section — ANN bench skipped)");
+            return Ok(());
+        }
+        let ann_times = run_bench(&runtime, &queries, |rt, q| rt.search_ann(q, k, ef))?;
+        println!("ANN ef={} ({} queries):", ef, n_queries);
+        print_latency(&ann_times);
+
+        // Recall@k of ANN vs exact, computed on the same queries.
+        let mut hits_overlap_total = 0.0f64;
+        for q in &queries {
+            let exact = runtime.search(q, k)?;
+            let approx = runtime.search_ann(q, k, ef)?;
+            let exact_set: std::collections::HashSet<&str> =
+                exact.hits.iter().map(|h| h.chunk_id.as_str()).collect();
+            let overlap = approx
+                .hits
+                .iter()
+                .filter(|h| exact_set.contains(h.chunk_id.as_str()))
+                .count();
+            hits_overlap_total += overlap as f64 / k as f64;
+        }
+        println!(
+            "  recall@{} (ANN vs exact): {:.4}",
+            k,
+            hits_overlap_total / n_queries as f64
+        );
+    }
+    Ok(())
+}
+
+fn run_bench(
+    rt: &nest_runtime::MmapNestFile,
+    queries: &[Vec<f32>],
+    mut f: impl FnMut(
+        &nest_runtime::MmapNestFile,
+        &[f32],
+    ) -> Result<nest_runtime::SearchResult, nest_runtime::RuntimeError>,
+) -> Result<Vec<f64>> {
+    let mut times = Vec::with_capacity(queries.len());
+    for q in queries {
+        let t0 = std::time::Instant::now();
+        f(rt, q)?;
+        times.push(t0.elapsed().as_secs_f64() * 1000.0);
+    }
+    times.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
+    Ok(times)
+}
+
+fn print_latency(times: &[f64]) {
+    let p = |q: f64| -> f64 {
+        let idx = ((times.len() as f64 - 1.0) * q).round() as usize;
+        times[idx]
+    };
+    let mean = times.iter().sum::<f64>() / times.len() as f64;
+    println!("  mean:   {:.3} ms", mean);
+    println!("  p50:    {:.3} ms", p(0.50));
+    println!("  p95:    {:.3} ms", p(0.95));
+    println!("  p99:    {:.3} ms", p(0.99));
+}
